Introduce nav controller
* Introduce a NavHost at the top level of the app that drives transitions
* Use deeplink-style handling of implicit VIEW intent
* Organize screens a bit more
